requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-generativeai==0.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
